fix(pgTextChunks): pin ::int overload so SUBSTRING doesn't hit SIMILAR TO
The chunked BODY[] streaming SQL

  SELECT SUBSTRING(<col> FROM $3 FOR $4) FROM mails WHERE ...

throws `invalid escape string` from Postgres against the real prod DB
(and any real Postgres 16). Root cause: `pg` sends JS number params as
text with no OID hint, so Postgres has to infer types from the
SUBSTRING call site. Three overloads share the shape:

  substring(text, int, int)     — numeric offsets (what we want)
  substring(text, text, text)   — SIMILAR TO pattern + escape char
  substring(text, text)         — regex pattern

Two-text params make Postgres pick overload 2, then read $4
(`"12000"`) as the ESCAPE CHARACTER — which must be exactly one
character — and throws. Every lazy-body BODY[] stream fails on the
first chunk. The client never assembles a response.

Explicit `$3::int FOR $4::int` casts pin the intended overload.
Verified against prod DB: without casts fails, with casts returns the
expected chunk.

Two mock tests hard-matched the pre-fix SQL shape; updated to match
`\$3::int\s+FOR\s+\$4::int` so a future accidental cast removal fails
the mock's chunk lookup → tests fail → regression caught.

Refs: #757

+++ src/server/lib/imap/fetch-helpers-lazy-body.test.ts
@@ -27,7 +27,7 @@ const substringCalls: Array<{ column: "text" | "html"; take: number; returnedCha
 const columnStore = new Map<string, string>();
 
 const mockQuery = mock(async (sql: string, values: unknown[]) => {
-  const substringMatch = sql.match(/SUBSTRING\((text|html)\s+FROM\s+\$3\s+FOR\s+\$4\)/);
+  const substringMatch = sql.match(/SUBSTRING\((text|html)\s+FROM\s+\$3::int\s+FOR\s+\$4::int\)/);
   if (substringMatch) {
     const column = substringMatch[1] as "text" | "html";
     const [mail_id, , offset, take] = values as [string, string, number, number];
