+++ src/server/lib/imap/session-utils-lazy-text.test.ts
@@ -24,7 +24,7 @@ const columnStore = new Map<string, string>();
 const mockQuery = mock(async (sql: string, values: unknown[]) => {
   // Route SUBSTRING(text FROM ... FOR ...) / SUBSTRING(html FROM ... FOR ...)
   // through the columnStore. Every non-SUBSTRING query returns empty rows.
-  const substringMatch = sql.match(/SUBSTRING\((text|html)\s+FROM\s+\$3\s+FOR\s+\$4\)/);
+  const substringMatch = sql.match(/SUBSTRING\((text|html)\s+FROM\s+\$3::int\s+FOR\s+\$4::int\)/);
   if (substringMatch) {
     const column = substringMatch[1] as "text" | "html";
     const [mail_id, , offset, take] = values as [string, string, number, number];
